Squashed 'rtengine/libraw/' changes from cccb97647..1ef70158d
1ef70158d 0.21.2 release
62f042366 tag type => tag size mapping fixed
ee087e3fe cubic_spline: better handling of non-integer data
af755b991 extra metadata check in arq_load_raw
0fadd8819 Better incorrect data handling in cubic_spline
d7fb66053 skip invalid pattern in xtrans_interpolate
d059ed280 Check HL recovery coeffs before processing
104730519 limit wavelet denoise minimum size
cae09838e raw-identify: use fallback if PATH_MAX not available
d6c677608 additional check against corrupted ljpeg layout
1001a6ac1 Disable color conversion for Canon 16-bit thumbnails
a5130b01b docs/changelog: explained the case when no thumbnail is found in specific file
600c0c63d rename swapXX to libraw_swapXX to avoid name conflict
299c8a11b Check against corrupted LJPEG header in Canon sRAW decoder
ec8671ad9 Limit embedded color profile allocation/read size
5229d5942 Wrong alloc result check for 16-bit bitmap thumbnail
b278b775f check pana_data/buffer offset before use
7f4b8d3af Check P1 quadrant linearization coeff[15] against zero
e942a7db6 avoid integer overflow in buffer space check
f6a57cfb8 prevent buffer overrun in buffer_datastream::scanf_one
3e62ed304 ensure correct T.tlength for 16b bitmap thumbnails(2)
8e52d81cd ensure correct T.tlength for 16b bitmap thumbnails
8e1af15e2 Do not run sraw decoder on (crafted) bayer files
0ace959c2 better striped thumbnails handling
477e0719f do not set shrink flag for 3/4 component images
c8efae6c5 allow more decoders for fuji-rotated RAWs

git-subtree-dir: rtengine/libraw
git-subtree-split: 1ef70158d7fde1ced6aaddb0b9443c32a7121d3d

+++ src/metadata/fuji.cpp
@@ -132,7 +132,7 @@ void LibRaw::parseAdobeRAFMakernote()
   }
 
 #define CHECKSPACE(s)                                                          \
-  if (posPrivateMknBuf + (s) > PrivateMknLength)                               \
+  if (INT64(posPrivateMknBuf) + INT64(s) > INT64(PrivateMknLength))            \
   {                                                                            \
     free(PrivateMknBuf);                                                       \
     return;                                                                    \
@@ -209,7 +209,7 @@ void LibRaw::parseAdobeRAFMakernote()
     PrivateOrder = sget2(PrivateMknBuf);
     unsigned s, l;
     s = ifd_start = sget4(PrivateMknBuf +2)+6;
-    CHECKSPACE(ifd_start+4);
+    CHECKSPACE(INT64(ifd_start)+4LL);
     l = ifd_len = sget4(PrivateMknBuf +ifd_start);
 	CHECKSPACE_ABS3(ifd_start, ifd_len, 4);
 
@@ -767,7 +767,7 @@ void LibRaw::parseAdobeRAFMakernote()
 
 		  if (wb_section_offset)
 		  {
-			  CHECKSPACE(wb_section_offset + 12);
+			  CHECKSPACE(INT64(wb_section_offset) + 12LL);
 		  }
 
           if (wb_section_offset &&
